moved almost everything of the battery into the Battery.swift file.
